+++ routes/noticesRoute.js
@@ -13,6 +13,7 @@ const {
 const ctrlWrapper = require('../helpers/ctrlWrapper');
 const authMdw = require('../middlewares/authMdw');
 const avatarStorrageMdw = require('../middlewares/avatarStorrageMdw');
+const noticeValidationMdw = require('../middlewares/noticeValidationMdw');
 
 const router = express.Router();
 
@@ -22,7 +23,7 @@ router.delete('/favoriteads/:id', authMdw, ctrlWrapper(delAuthFavNotice));
 router.get('/myads', authMdw, ctrlWrapper(getAuthOwnNotice));
 router.delete('/myads/:id', authMdw, ctrlWrapper(delAuthOwnNotice));
 
-router.post('/new', authMdw, avatarStorrageMdw, ctrlWrapper(addNotice));
+router.post('/new', authMdw, noticeValidationMdw, avatarStorrageMdw, ctrlWrapper(addNotice));
 router.get('/:id', ctrlWrapper(getNotice));
 router.get('/category/:category', ctrlWrapper(getNoticeByCategory));
